moved `EVForm` to `ev_form.dart` +  code cleanup
